int to long (Field)

+++ saarsiedler-group16/src/main/java/de/unisaarland/cs/sopra/common/model/Field.java
@@ -27,7 +27,7 @@ public class Field {
 		return this.fieldType;
 	}
 	
-	public int getNumber() {
+	public long getNumber() {
 		return this.number;
 	}
